% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol_tax.R
\docType{data}
\name{alcohol_tax}
\alias{alcohol_tax}
\title{A simulated dataset based on the epi.bfi dataset from the psychTools package}
\format{
A data frame with 3666 rows and 5 variables:
\describe{
  \item{alcohol}{Number of standard drinks consumed in a month}
  \item{Year}{Year when the data was collected}
  \item{Postcode}{Postcode of the participant}
  \item{State}{State}
  \item{Remoteness}{Captial city or regional area}

}
}
\usage{
alcohol_tax
}
\description{
Alcohol consumption data from 3666 participants
}
\keyword{datasets}
