% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{A simulated dataset based on the simulated cancer dataset from UCLA Institute for
Digital Research and Education}
\format{
A data frame with 5959 rows and 12 variables:
\describe{
  \item{pain}{Pain level ranged from 1 to 9}
  \item{remission}{1: Remitted; 0: Not remitted}
  \item{Age}{Patient age}
  \item{SmokingHx}{Smoking status: former smoker/ current smoker/ non-smoker}
  \item{Sex}{Patient sex: male/ female}
  \item{CancerStage}{Cancer Stage: I/ II/ III/ IV}
  \item{LengthofStay}{Length of hospital stay}
  \item{BMI}{Body Mass Index}
  \item{DID}{Doctor ID}
  \item{Experience}{Doctor's years of experience}
  \item{HID}{Hospital ID}
  \item{TumourSize}{Tumor Size}

}
}
\source{
\url{https://stats.idre.ucla.edu/r/dae/mixed-effects-logistic-regression/}
}
\usage{
cancer
}
\description{
Lung cancer data from 5959 patients
}
\keyword{datasets}
