% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{Life expectancy, population and GDP data from the gapminder package}
\format{
A data frame with 1704 observations.
\describe{
  \item{Reaction}{Number of standard drinks consumed in a month}
  \item{Days}{Year when the data was collected}
  \item{Subject}{Subject number on which the observation was made.}

}
}
\usage{
gapminder
}
\description{
This dataset is sourced from the gapminder package, originally named 'sleepstudy'.
}
\keyword{datasets}
