% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDI.R
\docType{data}
\name{HDI}
\alias{HDI}
\title{Human Development Index 1990-2018 by regions}
\format{
A data frame with 199 rows and 11 variables:
\describe{
  \item{Region}{Region}
  \item{HDI_1990}{Human Development Index 1990}
  \item{GDP_1990}{GDP per capita 1990}

}
}
\source{
Data were adapted from the epi.bfi dataset from the psychTools package
}
\usage{
HDI
}
\description{
Human Development Index by regions from 1990 to 2018 in wide format
}
\keyword{datasets}
