% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personality.R
\docType{data}
\name{personality}
\alias{personality}
\title{A simulated dataset based on the epi.bfi dataset from the psychTools package}
\format{
A data frame with 231 rows and 9 variables:
\describe{
  \item{Agreeableness}{Big 5 personality trait - Agreeableness}
  \item{Conscientiousness}{Big 5 personality trait - Conscientiousness}
  \item{Extraversion}{Big 5 personality trait - Extraversion}
  \item{Neuroticism}{Big 5 personality trait - Neuroticism}
  \item{Openness}{Big 5 personality trait - Openness}
  \item{Depression}{Beck Depression scale}
  \item{Trait Anxiety}{Trait Anxiety}
  \item{State Anxiety}{State Anxiety}

}
}
\source{
Data were adapted from the epi.bfi dataset from the psychTools package
}
\usage{
personality
}
\description{
Big 5 persoanlity (Agreeableness, Conscientiousness, Extraversion,
Neuroticism, Openness), Depression, Anxiety and Sex
}
\keyword{datasets}
