% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\docType{data}
\name{sleep}
\alias{sleep}
\title{Reaction times in sleep deprivation study}
\format{
A data frame with 180 observations.
\describe{
  \item{Reaction}{Number of standard drinks consumed in a month}
  \item{Days}{Year when the data was collected}
  \item{Subject}{Subject number on which the observation was made.}

}
}
\usage{
sleep
}
\description{
This dataset is sourced from the lme4 package, originally named 'sleepstudy'.
The average reaction time per day for subjects in a sleep deprivation study.
On day 0 the subjects had their normal amount of sleep.
Starting that night they were restricted to 3 hours of sleep per night.
The observations represent the average reaction time on a series of tests given each day to each subject.
}
\keyword{datasets}
