% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\docType{data}
\name{stars}
\alias{stars}
\title{A dataset of star temperature and light intensity}
\format{
A data frame 47 observations and 2 variabless:
\describe{
  \item{temperature}{Surface temperature of a star}
  \item{light}{light intensity}

}
}
\usage{
stars
}
\description{
This dataset is sourced from the robustbase package, originally named StarsCYG.
It contains 47 observations of star temperature and light intensity.
}
\keyword{datasets}
