% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_server.R
\name{start_server}
\alias{start_server}
\title{Starting the R server for StatsNotebook
This function starts the R server. Future implementation will allow change of port
Use "off" to turn off server}
\usage{
start_server()
}
\description{
Starting the R server for StatsNotebook
This function starts the R server. Future implementation will allow change of port
Use "off" to turn off server
}
