% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substance.R
\docType{data}
\name{substance}
\alias{substance}
\title{A simulated dataset on youth substance misuse and family intervention}
\format{
A data frame with 231 rows and 9 variables:
\describe{
  \item{gender}{Female/ Male}
  \item{conflict}{level of family conflict at Grade 7}
  \item{dev_peer}{Involvement in deviant peer group at Grade 8}
  \item{sub_exp}{Experimentation of substance at Grade 8}
  \item{fam_int}{Participation in family intervention program at junior high school at Grade 7}
  \item{sub_disorder}{Substance use disorder in young adulthood}

}
}
\usage{
substance
}
\description{
A simulated dataset of 553 participants collected at Grade 7, Grade 8 and age 25.
}
\keyword{datasets}
