% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNDP.R
\docType{data}
\name{UNDP}
\alias{UNDP}
\title{A Dataset from the United Nations Development Programme}
\format{
A data frame with 199 rows and 11 variables:
\describe{
  \item{HDI_Rank_2018}{Human Development Index Ranking 2018}
  \item{Country}{Country name}
  \item{HDI}{Human Development Index 2018}
  \item{Pop}{Population (millions) in 2018}
  \item{GII}{Gender Inequality Index in 2018}
  \item{CO2}{Crabon dioxide emissions, per capita (tonnes)}
  \item{Prison}{Prison population (per 100,000 people) in 2017}
  \item{Homocide}{Crabon dioxide emissions, per capita (tonnes)}
  \item{GDP}{Crabon dioxide emissions, per capita (tonnes)}
  \item{Schooling}{Crabon dioxide emissions, per capita (tonnes)}
  \item{Continent}{Continent}

}
}
\usage{
UNDP
}
\description{
A dataset of 199 countries
}
\keyword{datasets}
