% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellbeing.R
\docType{data}
\name{wellbeing}
\alias{wellbeing}
\title{A simulated dataset on wellbeing, stress and social support}
\format{
A data frame with 231 rows and 9 variables:
\describe{
  \item{stress}{level of stress}
  \item{SocialSupport}{Level of social support}
  \item{Wellbeing}{Level of psychological wellbeing}
  \item{Sex}{Male/ Female}

}
}
\usage{
wellbeing
}
\description{
A simulated dataset of 1000 participants
}
\keyword{datasets}
